pyproject.toml
[project]
name = "situation-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
